Remove WooCommerce as clone of Shopify
WooCommerce is not a clone of Shopify. It's an open-source WordPress plugin: https://github.com/woocommerce/woocommerce Shopify is a proprietary, hosted platform. The only sense in which you could say it's a "clone" is that both are for the purpose of e-commerce, but that doesn't match the common usage of the word "clone" IMHO.